+++ frontend/src/pages/Contracts/Contract/ContractDetails/InvoiceSection/fields.tsx
@@ -1,0 +1,29 @@
+import { FormikValues } from "formik";
+import { UseQueryResult } from "react-query";
+import { IEditFields } from "types";
+
+/**
+ * The view fields.
+ *
+ * @param   {UseQueryResult<FormikValues>} query The react query data for specific project.
+ * @returns {Array}
+ */
+export const readFields = (query: UseQueryResult<FormikValues>) => {
+  return [{ width: "half", title: "Fiscal Year", value: query?.data?.data?.fiscal_year }];
+};
+
+/**
+ * The edit fields.
+ *
+ * @returns {Array}
+ */
+export const editFields: () => IEditFields[] = () => {
+  return [
+    {
+      fieldName: "fiscal_year",
+      fieldType: "singleText",
+      fieldLabel: "Fiscal Year",
+      width: "half",
+    },
+  ];
+};